requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
ta>=0.10.0
yfinance>=0.2.0
pandas-datareader>=0.10.0
streamlit>=1.30.0
plotly>=5.18.0
joblib>=1.3.0
lightgbm>=4.0.0
shap>=0.40.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
